requirements.txt
Flask~=2.0.1
psycopg2-binary==2.9.1
gunicorn
pandas=1.3.4
lightgbm=3.3.1
xgboost=1.5.0
catboost=1.0.1
imblearn~=0.0
pandas~=1.3.4
numpy~=1.21.3
lightgbm~=3.3.1
joblib~=1.1.0
xgboost~=1.5.0
catboost~=1.0.1
scikit-learn~=1.0.1